Fix - Input macros working weird
The display just broke if the screen was small OR the text too long, and the logic just didn't work if there were multiple input of the same type in the entry. It is ideal to have constant size inputs? No, but that's a fix for a less stupid future me.